Add basic "join" button
- Add "join" button that enables one-click event joining
- When the user selects an event to display it in the window, the whole event is downloaded. This is a preparation for not downloading all event details when only id and name are needed.

+++ shared/src/main/scala/scalac/octopusonwire/shared/domain/Event.scala
@@ -1,3 +1,3 @@
 package scalac.octopusonwire.shared.domain
 
-case class Event(id: Int, name: String, startDate: Long, endDate: Long, location: String, url: String)
+case class Event(id: Long, name: String, startDate: Long, endDate: Long, location: String, url: String)
